Move the most long-running kafka tests to a separate batch
